[refs] Split core table+card metadata functions into new and old
Begin the process of forking the library to work with new
`:ident`-powered refs, and new metadata formats.

`foo:old-refs` is the new name for the original code.
`foo:new-refs` is the new code, but some of them are stubs that throw,
in this PR.

New `lib.tu` helper macros for writing tests that explicitly target old,
new, or both forms.